// generated-api/models/CategoryWeightResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Nexus Boxing API
 * REST API for authentication, all time ranked boxers, perfect boxer generation, updating category weights, making predictions, managing prediction histories, and weight classes.
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * Category weight response
 * @export
 * @interface CategoryWeightResponse
 */
export interface CategoryWeightResponse {
    /**
     * Weight class ID
     * @type {number}
     * @memberof CategoryWeightResponse
     */
    weightClassId?: number;
    /**
     * Physical category weight
     * @type {number}
     * @memberof CategoryWeightResponse
     */
    physicalWeight?: number;
    /**
     * Technical category weight
     * @type {number}
     * @memberof CategoryWeightResponse
     */
    technicalWeight?: number;
    /**
     * Tactical category weight
     * @type {number}
     * @memberof CategoryWeightResponse
     */
    tacticalWeight?: number;
    /**
     * Psychological category weight
     * @type {number}
     * @memberof CategoryWeightResponse
     */
    psychologicalWeight?: number;
    /**
     * Experience category weight
     * @type {number}
     * @memberof CategoryWeightResponse
     */
    experienceWeight?: number;
}

/**
 * Check if a given object implements the CategoryWeightResponse interface.
 */
export function instanceOfCategoryWeightResponse(value: object): value is CategoryWeightResponse {
    return true;
}

export function CategoryWeightResponseFromJSON(json: any): CategoryWeightResponse {
    return CategoryWeightResponseFromJSONTyped(json, false);
}

export function CategoryWeightResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): CategoryWeightResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'weightClassId': json['weightClassId'] == null ? undefined : json['weightClassId'],
        'physicalWeight': json['physicalWeight'] == null ? undefined : json['physicalWeight'],
        'technicalWeight': json['technicalWeight'] == null ? undefined : json['technicalWeight'],
        'tacticalWeight': json['tacticalWeight'] == null ? undefined : json['tacticalWeight'],
        'psychologicalWeight': json['psychologicalWeight'] == null ? undefined : json['psychologicalWeight'],
        'experienceWeight': json['experienceWeight'] == null ? undefined : json['experienceWeight'],
    };
}

export function CategoryWeightResponseToJSON(json: any): CategoryWeightResponse {
    return CategoryWeightResponseToJSONTyped(json, false);
}

export function CategoryWeightResponseToJSONTyped(value?: CategoryWeightResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'weightClassId': value['weightClassId'],
        'physicalWeight': value['physicalWeight'],
        'technicalWeight': value['technicalWeight'],
        'tacticalWeight': value['tacticalWeight'],
        'psychologicalWeight': value['psychologicalWeight'],
        'experienceWeight': value['experienceWeight'],
    };
}