// generated-api/models/CategoryWeightUpdateRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Nexus Boxing API
 * REST API for authentication, all time ranked boxers, perfect boxer generation, updating category weights, making predictions, managing prediction histories, and weight classes.
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface CategoryWeightUpdateRequest
 */
export interface CategoryWeightUpdateRequest {
    /**
     * 
     * @type {number}
     * @memberof CategoryWeightUpdateRequest
     */
    physicalWeight: number;
    /**
     * 
     * @type {number}
     * @memberof CategoryWeightUpdateRequest
     */
    technicalWeight: number;
    /**
     * 
     * @type {number}
     * @memberof CategoryWeightUpdateRequest
     */
    tacticalWeight: number;
    /**
     * 
     * @type {number}
     * @memberof CategoryWeightUpdateRequest
     */
    psychologicalWeight: number;
    /**
     * 
     * @type {number}
     * @memberof CategoryWeightUpdateRequest
     */
    experienceWeight: number;
}

/**
 * Check if a given object implements the CategoryWeightUpdateRequest interface.
 */
export function instanceOfCategoryWeightUpdateRequest(value: object): value is CategoryWeightUpdateRequest {
    if (!('physicalWeight' in value) || value['physicalWeight'] === undefined) return false;
    if (!('technicalWeight' in value) || value['technicalWeight'] === undefined) return false;
    if (!('tacticalWeight' in value) || value['tacticalWeight'] === undefined) return false;
    if (!('psychologicalWeight' in value) || value['psychologicalWeight'] === undefined) return false;
    if (!('experienceWeight' in value) || value['experienceWeight'] === undefined) return false;
    return true;
}

export function CategoryWeightUpdateRequestFromJSON(json: any): CategoryWeightUpdateRequest {
    return CategoryWeightUpdateRequestFromJSONTyped(json, false);
}

export function CategoryWeightUpdateRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): CategoryWeightUpdateRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'physicalWeight': json['physicalWeight'],
        'technicalWeight': json['technicalWeight'],
        'tacticalWeight': json['tacticalWeight'],
        'psychologicalWeight': json['psychologicalWeight'],
        'experienceWeight': json['experienceWeight'],
    };
}

export function CategoryWeightUpdateRequestToJSON(json: any): CategoryWeightUpdateRequest {
    return CategoryWeightUpdateRequestToJSONTyped(json, false);
}

export function CategoryWeightUpdateRequestToJSONTyped(value?: CategoryWeightUpdateRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'physicalWeight': value['physicalWeight'],
        'technicalWeight': value['technicalWeight'],
        'tacticalWeight': value['tacticalWeight'],
        'psychologicalWeight': value['psychologicalWeight'],
        'experienceWeight': value['experienceWeight'],
    };
}